requirements.txt
Django==4.2.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
docker==7.1.0
docker-py==1.10.6
docker-pycreds==0.4.0
celery==5.4.0
jsonpath==0.82.2
jsonschema==4.22.0
paramiko==3.4.0
pillow==10.4.0
pipreqs==0.5.0
psycopg==3.1.18
psycopg2==2.9.9
requests==2.31.0
networkx==3.3
matplotlib==3.9.2
